import_data/requirements.txt
weaviate-client===4.18.3
datasets==4.4.1
